Add test case for handling '-' which ends up being NaN in stats
The git log reports '-' under insertions/additions and deletions for binary files. This ends up being NaN due to the math operations mapped to it, which according to Section 2.4 of JSON specification special double values are not allowed (NaN, Infinity, -Infinity).

This "issue", not sure if you think it's an issue actually haha, is shadowed/masked by the fact that JSON.stringify() will convert the NaN to null. I only noticed it while making a quick script that creats .mock and .expected for test cases, so I could get a mock for the maxbuffer issue.

This commit adds test cases for this.