Migrate git2json from being exec() based to being spawn()based
This fixes the stdout maxBuffer exceeded issue  (#1) that prevented the use of git2json on any repository with git log output exceeding 200KB, which was the default for child_process.exec(). Rather than upping the default buffer size of child_process.exec() to a size that MIGHT accommodate MOST of the large repos, I felt the correct decision was to replace child_process.exec() with child_process.spawn.

Luckily it was rather straight forward. Only one additional change that adapts how the command/args are passed to the respective functions accompanies the bulk changes that substitute spawn() for exec().